Config revamp / limit to txt
Program now prompts for input using PyQt5 rather than command prompt.
Renamed file to .pyw so that the command prompt window doesn't open and
the file runs only in the system tray.

Changed the list_files function so that it now only adds .txt files to
the menu.

Updated README